Move more early buffered lints for `rustc_resolve` to dyn lint diagnostics
